+++ contracts/NFTFloorPriceConsumerV3.sol
@@ -4,7 +4,6 @@ pragma solidity ^0.8.7;
 import "@chainlink/contracts/src/v0.8/interfaces/AggregatorV3Interface.sol";
 
 contract NFTFloorPriceConsumerV3 {
-
     AggregatorV3Interface internal nftFloorPriceFeed;
 
     /**
@@ -13,19 +12,22 @@ contract NFTFloorPriceConsumerV3 {
      * Address: 0x5c13b249846540F81c093Bc342b5d963a7518145
      */
     constructor() {
-        nftFloorPriceFeed = AggregatorV3Interface(0x5c13b249846540F81c093Bc342b5d963a7518145);
+        nftFloorPriceFeed = AggregatorV3Interface(
+            0x5c13b249846540F81c093Bc342b5d963a7518145
+        );
     }
 
     /**
      * Returns the latest price
      */
-    function getLatestPrice() public view returns (int) {
+    function getLatestPrice() public view returns (int256) {
         (
-            /*uint80 roundID*/,
-            int nftFloorPrice,
-            /*uint startedAt*/,
-            /*uint timeStamp*/,
-            /*uint80 answeredInRound*/
+            ,
+            /*uint80 roundID*/
+            int256 nftFloorPrice, /*uint startedAt*/ /*uint timeStamp*/ /*uint80 answeredInRound*/
+            ,
+            ,
+
         ) = nftFloorPriceFeed.latestRoundData();
         return nftFloorPrice;
     }
