+++ contracts/ProofOfReserveConsumerV3.sol
@@ -4,7 +4,6 @@ pragma solidity ^0.8.7;
 import "@chainlink/contracts/src/v0.8/interfaces/AggregatorV3Interface.sol";
 
 contract ReserveConsumerV3 {
-
     AggregatorV3Interface internal reserveFeed;
 
     /**
@@ -13,19 +12,22 @@ contract ReserveConsumerV3 {
      * Address: 0xa81FE04086865e63E12dD3776978E49DEEa2ea4e
      */
     constructor() {
-        reserveFeed = AggregatorV3Interface(0xa81FE04086865e63E12dD3776978E49DEEa2ea4e);
+        reserveFeed = AggregatorV3Interface(
+            0xa81FE04086865e63E12dD3776978E49DEEa2ea4e
+        );
     }
 
     /**
      * Returns the latest price
      */
-    function getLatestReserve() public view returns (int) {
+    function getLatestReserve() public view returns (int256) {
         (
-            /*uint80 roundID*/,
-            int reserve,
-            /*uint startedAt*/,
-            /*uint timeStamp*/,
-            /*uint80 answeredInRound*/
+            ,
+            /*uint80 roundID*/
+            int256 reserve, /*uint startedAt*/ /*uint timeStamp*/ /*uint80 answeredInRound*/
+            ,
+            ,
+
         ) = reserveFeed.latestRoundData();
         return reserve;
     }
